Move `src/game/bezier.cpp/h` to `src/base/bezier.cpp/h`
As those functions are not game specific per se, they can be part of base instead.

Closes #3334.
